feat: add selectUnderlyingLpAssetsWithBalancesAndIcons
